src: fix memory leaks using TACO runtime API from reference cycle
This commit fixes a memory leak in the TACO runtime that doesn't
allow for reclamation of memory used when tensors are allocated
by the TACO runtime. In particular, when accessing a tensor, the
resulting `Access` took an RC-pointer to the tensor, and then tensor
then stored this access within itself, causing a cycle.